fix(cypress-okta-ete): split registration tests into 2 files as they were taking too long to run in a single file
