+++ .github/workflows/cypress-ete-okta.yml
@@ -9,7 +9,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        group: [1, 2, 3, 4, 5]
+        group: [1, 2, 3, 4, 5, 6]
     timeout-minutes: 15
     services:
       redis:
